requirements.txt
decorator==4.1.2
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
jsonschema==2.6.0
kafka==1.3.5
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyaml==17.12.1
Pygments==2.2.0
python-dateutil==2.6.1
PyYAML==3.12
simplegeneric==0.8.1
six==1.11.0
strict-rfc3339==0.7
traitlets==4.3.2
wcwidth==0.1.7
